requirements.txt
fuzzywuzzy==0.18.0
openai==1.66.3
PyYAML==6.0.2
requests==2.32.3